Lighthouse Optimisations
Made some minor improvements to the Performance (as per Lighthouse's recommendations).

Get a minor increase by using Preact, minimised the avatar image but that didn't seem to do much. Removed some non-standard CSS in the template that's only useful for IE8 (which I think is long dead, doubt the people i'll be showing this to will be too bothered.)